[build] Inhibit spurious array bounds warning on some versions of gcc
Some versions of gcc (observed with gcc 9.3.0 on NixOS Linux) produce
a spurious warning about an out-of-bounds array access for the
isa_extra_probe_addrs[] array.

Work around this compiler bug by redefining the array index as a
signed long, which seems to somehow avoid this spurious warning.

Debugged-by: Manuel Mendez <mmendez534@gmail.com>

+++ src/drivers/bus/isa.c
@@ -95,7 +95,7 @@ static void isa_remove ( struct isa_device *isa ) {
 static int isabus_probe ( struct root_device *rootdev ) {
 	struct isa_device *isa = NULL;
 	struct isa_driver *driver;
-	int ioidx;
+	long ioidx;
 	int rc;
 
 	for_each_table_entry ( driver, ISA_DRIVERS ) {
